ecommerce/backend/requirements.txt
FastAPI>=0.104.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
uvicorn>=0.24.0
httpx>=0.25.0
PyJWT>=2.8.0
requests>=2.31.0
cryptography>=42.0.0

